Major refactor of classes
